+++ src/test/java/org/ld4l/bib2lod/entitybuilders/marcxml/ld4l/AdminMetadataBuilderTest.java
@@ -4,16 +4,19 @@ import static org.ld4l.bib2lod.testing.xml.testrecord.MockMarcxml.MINIMAL_RECORD
 
 import org.junit.Assert;
 import org.junit.Before;
+import org.junit.BeforeClass;
 import org.junit.Test;
 import org.ld4l.bib2lod.datatypes.XsdDatatype;
 import org.ld4l.bib2lod.entity.Entity;
 import org.ld4l.bib2lod.entity.InstanceEntity;
 import org.ld4l.bib2lod.entitybuilders.BuildParams;
 import org.ld4l.bib2lod.entitybuilders.EntityBuilder.EntityBuilderException;
+import org.ld4l.bib2lod.entitybuilders.EntityBuilderFactory;
 import org.ld4l.bib2lod.ontology.ld4l.Ld4lAgentType;
 import org.ld4l.bib2lod.ontology.ld4l.Ld4lDatatypeProp;
 import org.ld4l.bib2lod.ontology.ld4l.Ld4lObjectProp;
 import org.ld4l.bib2lod.testing.AbstractTestClass;
+import org.ld4l.bib2lod.testing.BaseMockBib2LodObjectFactory;
 import org.ld4l.bib2lod.testing.xml.testrecord.MockMarcxml;
 
 
@@ -40,14 +43,34 @@ public class AdminMetadataBuilderTest extends AbstractTestClass {
                  "</datafield>" +
             "</record>");
     
+    public static final MockMarcxml SOURCE_040$a = TEST_RECORD.openCopy()
+            .findDatafield("040")
+            .addSubfield("a", "NIC")
+            .deleteSubfield("c")
+            .lock();
+
+    
+    public static final MockMarcxml SOURCE_040$a$c = TEST_RECORD.openCopy()
+            .findDatafield("040")
+            .addSubfield("a", "NIC")
+            .lock();
+    
     public static final MockMarcxml INVALID_005_VALUE =  MINIMAL_RECORD.openCopy()
             .addControlfield("005", "20130330")
             .lock();
     
+    private static BaseMockBib2LodObjectFactory factory;
     private AdminMetadataBuilder builder;   
     
+    @BeforeClass
+    public static void setUpOnce() throws Exception {
+        factory = new BaseMockBib2LodObjectFactory();  
+        factory.addInstance(EntityBuilderFactory.class, 
+                new MarcxmlToLd4lEntityBuilderFactory());
+    }
+    
     @Before
-    public void setUp() {       
+    public void setUp() {     
         this.builder = new AdminMetadataBuilder();
     }
     
@@ -80,8 +103,8 @@ public class AdminMetadataBuilderTest extends AbstractTestClass {
     public void testInstanceHasAdminMetadata() throws Exception {
         Entity instance = new InstanceEntity();
         Entity adminMetadata = buildAdminMetadata(instance, TEST_RECORD);
-        Assert.assertSame(instance.getChild(Ld4lObjectProp.HAS_ADMIN_METADATA), 
-                adminMetadata);
+        Assert.assertSame(instance.getChild( 
+                Ld4lObjectProp.HAS_ADMIN_METADATA), adminMetadata);
     }
     
     @Test
@@ -92,34 +115,49 @@ public class AdminMetadataBuilderTest extends AbstractTestClass {
     }
     
     @Test
-    public void testHasSource() throws Exception {
-        Entity adminMetadata = buildAdminMetadata(TEST_RECORD);
-        Assert.assertNotNull(adminMetadata.getChild(Ld4lObjectProp.HAS_SOURCE));
+    public void testSource_040$a() throws Exception {
+        Entity adminMetadata = buildAdminMetadata(SOURCE_040$a);
+        Assert.assertNotNull(
+                adminMetadata.getChild(Ld4lObjectProp.HAS_SOURCE));
     }
     
     @Test
-    public void testSourceIsAgent() throws Exception {
+    public void testSource_040$c() throws Exception {
+        Entity adminMetadata = buildAdminMetadata(TEST_RECORD);
+        Assert.assertNotNull(
+                adminMetadata.getChild(Ld4lObjectProp.HAS_SOURCE));
+    }
+    
+    @Test
+    public void testTwoSources_040$a$c() throws Exception {
+        Entity adminMetadata = buildAdminMetadata(SOURCE_040$a$c);
+        Assert.assertEquals(2, adminMetadata.getChildren(
+                Ld4lObjectProp.HAS_SOURCE).size());
+    }
+    
+    @Test
+    public void testSourceIsAgent_040() throws Exception {
         Entity adminMetadata = buildAdminMetadata(TEST_RECORD);
         Entity agent = adminMetadata.getChild(Ld4lObjectProp.HAS_SOURCE);
         Assert.assertTrue(agent.hasType(Ld4lAgentType.superClass()));
     }
     
     @Test
-    public void testSourceHasName() throws Exception {
+    public void testSourceHasName_040() throws Exception {
         Entity adminMetadata = buildAdminMetadata(TEST_RECORD);
         Entity agent = adminMetadata.getChild(Ld4lObjectProp.HAS_SOURCE);
         Assert.assertEquals("NIC", agent.getValue(Ld4lDatatypeProp.NAME));
     }
     
     @Test 
-    public void testDescriptionModifiers() throws Exception {
+    public void testDescriptionModifiers_040$d() throws Exception {
         Entity adminMetadata = buildAdminMetadata(TEST_RECORD);
         Assert.assertEquals(2, (adminMetadata.getChildren(
                 Ld4lObjectProp.HAS_DESCRIPTION_MODIFIER)).size());        
     }
     
     @Test
-    public void testDescriptionModifierIsAgent() throws Exception {
+    public void testDescriptionModifierIsAgent_040$d() throws Exception {
         Entity adminMetadata = buildAdminMetadata(TEST_RECORD);  
         Entity agent = (adminMetadata.getChildren(
                 Ld4lObjectProp.HAS_DESCRIPTION_MODIFIER)).get(0);
@@ -127,7 +165,7 @@ public class AdminMetadataBuilderTest extends AbstractTestClass {
     }
     
     @Test
-    public void testDescriptionModifierHasName() throws Exception {
+    public void testDescriptionModifierHasName_040$d() throws Exception {
         Entity adminMetadata = buildAdminMetadata(TEST_RECORD);
         Entity agent = (adminMetadata.getChildren(
                 Ld4lObjectProp.HAS_DESCRIPTION_MODIFIER)).get(0);
@@ -135,16 +173,14 @@ public class AdminMetadataBuilderTest extends AbstractTestClass {
     }
     
     @Test
-    public void testDescriptionConventions() throws Exception {
+    public void testDescriptionConventions_040$e() throws Exception {
         Entity adminMetadata = buildAdminMetadata(TEST_RECORD);
-        Entity conventions = adminMetadata.getChild(
-                Ld4lObjectProp.HAS_DESCRIPTION_CONVENTIONS);
         Assert.assertEquals(2, (adminMetadata.getChildren(
                 Ld4lObjectProp.HAS_DESCRIPTION_CONVENTIONS)).size());   
     }
     
     @Test
-    public void testDescriptionConventionsValue() throws Exception {
+    public void testDescriptionConventionsValue_040$e() throws Exception {
         Entity adminMetadata = buildAdminMetadata(TEST_RECORD);
         Entity conventions = (adminMetadata.getChildren(
                 Ld4lObjectProp.HAS_DESCRIPTION_CONVENTIONS)).get(0);
@@ -152,21 +188,21 @@ public class AdminMetadataBuilderTest extends AbstractTestClass {
     }
     
     @Test
-    public void test005DateTimeValue() throws Exception {
+    public void testDateTimeValue_005() throws Exception {
         Entity adminMetadata = buildAdminMetadata(TEST_RECORD);
         Assert.assertEquals("2013-03-30T14:56:47", adminMetadata.getValue(
                 Ld4lDatatypeProp.CHANGE_DATE));
     }
     
     @Test
-    public void test005DateTimeDatatype() throws Exception {
+    public void testDateTimeDatatype_005() throws Exception {
         Entity adminMetadata = buildAdminMetadata(TEST_RECORD);
         Assert.assertSame(XsdDatatype.DATETIME, adminMetadata.getAttribute(
                 Ld4lDatatypeProp.CHANGE_DATE).getDatatype());
     }
     
     @Test
-    public void invalid005DateTimeValue_ThrowsException() throws Exception {
+    public void invalidDateTimeValue_005_ThrowsException() throws Exception {
         buildAndExpectException(INVALID_005_VALUE, 
                 "Invalid value for control field 005");
     }